webapp/requirements.txt
fastapi==0.110.3
uvicorn==0.29.0
pillow==10.3.0
python-multipart==0.0.9